requirements.txt
streamlit
Pillow
tensorflow
numpy
opencv-python
keras
matplotlib